remove shop and fix models and requests

+++ src/resolvers/mutations/records.ts
@@ -6,15 +6,20 @@ export async function addRecord(_: any, { data }: any, { currentUser }: any) {
   const newRecord = new RecordModel({
     ...data,
     date_added: new Date(),
-    // shop: currentUser.shop,
+    user: currentUser.id,
   });
   await newRecord.save();
   return newRecord;
 }
 
-export async function editRecord(_: any, { id, data }: any) {
+export async function editRecord(
+  _: any,
+  { id, data }: any,
+  { currentUser }: any
+) {
   const record = await RecordModel.findById(id);
   if (!record) throw new Error('Record does not exist');
+  if (record.user != currentUser.id) throw new Error('No permission to edit');
   await RecordModel.updateOne({ _id: new mongoose.Types.ObjectId(id) }, data);
   return await RecordModel.findById(id);
 }
@@ -39,9 +44,14 @@ export async function setSold(_: any, { id }: { id: string }) {
   return await RecordModel.findById(id);
 }
 
-export async function removeRecord(_: any, { id }: { id: string }) {
+export async function removeRecord(
+  _: any,
+  { id }: { id: string },
+  { currentUser }: any
+) {
   const record = await RecordModel.findById(id);
   if (!record) throw new Error('Record does not exist');
+  if (record.user != currentUser.id) throw new Error('No permission to edit');
   return RecordModel.findByIdAndDelete(id)
     .exec()
     .then((res: any) => {
